MSI-757: Add Inventory Index invalidation when Source is enabled or disabled.
